fixed for cycle for non-8 byte primitives, but became insane :)

+++ test_files/for_loop_tests/all_fields.c
@@ -1,6 +1,6 @@
 int main() {
-    int y = 0;
-    for(int x = 0; x < 10; x = x + 1) {
+    short y = 0;
+    for(short x = 0; x < 10; x = x + 1) {
         y = y + 1;
     }
     return y;
